ivec3 for normals

+++ MinecraftClone/ChunkVertex.h
@@ -6,11 +6,11 @@
 struct ChunkVertex {
 public:
 	glm::ivec3 position{};
-	glm::vec3 normal{};
+	glm::ivec3 normal{};
 	glm::vec2 texUv{};
 
 	ChunkVertex() = default;
-	ChunkVertex(const glm::ivec3& position, const glm::vec3& normal)
+	ChunkVertex(const glm::ivec3& position, const glm::ivec3& normal)
 		: position{ position }, normal{ normal } { }
 
 };
